Basic Homepage
within this version of the homepage the basic functionality has been
created. The homepage contains a list of each element needed in the
navigation which will be created in the next stage of the assignment. It
also contains a footer with quick function for the user to use such as
call, home, back and search.