Menu with collapsible items to display recipe
within the menu.html i implemented collapsible items and broke up the
menu into appetizers, main, beverages and dessert. This makes it easier
for the viewer to view all the parts of the menu in one place while also
being able to click on the food/drink item to see what ingredients it
had within it.