Declutter the restaurant Menu
i have erased the menu all in one idea due to thinking users may not be
aware of scrolling down to see there is more than what they can see on
there mobile or browser. This way of adding a list view with buttons to
external pages breaks up and declutters the menu. It also will make use
of the back button more.