Navigation enhanced feature
within the navigation page, i have added the google maps direction
service jquery preset. This script allows the user to input their
address and then click directions to automitically calculate directions
provided below and seen on the maps. This is a feature that makes it
easier and more conveinient for users to find the restaurant. The logo
was also updated